AS3 support + custom dialect support
If you are using AS3, you would require to use the AS3 client salt which is different to the AS2, so it is now available to be modified in Configuration.js. The same with the dialect of the database connection, you may change it from mysql to mariadb if you are using that.